ai-services/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
torch==2.5.1
torchvision==0.20.1
numpy==2.2.0
scikit-learn==1.5.2
rapidfuzz==3.10.1
joblib==1.5.2
Pillow==12.1.0
